Restructured the app component in preparation to add Redux. Added a little more backend functionality
