--- FirebaseConfig.js
@@ -1,25 +1,0 @@
-// Import the functions you need from the SDKs you need
-import { initializeApp } from "firebase/app";
-import { getAuth  } from "@firebase/auth";
-
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
-
-
-// Your web app's Firebase configuration
-const firebaseConfig = {
-  apiKey: "AIzaSyAq3F_9cn_1tJnNP1S3Soao4TCc79dOwaE",
-  authDomain: "mind-bridge-a93eb.firebaseapp.com",
-  projectId: "mind-bridge-a93eb",
-  storageBucket: "mind-bridge-a93eb.appspot.com",
-  messagingSenderId: "1052298113718",
-  appId: "1:1052298113718:web:2abbf8ed75b0c7e6f7ac17"
-};
-
-// Initialize Firebase
-//const app = initializeApp(firebaseConfig);
-
-export const FIREBASE_APP = initializeApp(firebaseConfig); 
-export const FIREBASE_AUTH = getAuth(FIREBASE_APP); 
-
-
