+++ components/auth/navigation/screens/Setting.js
@@ -1,9 +1,10 @@
 //Avlokita's Work
 
 import * as React from 'react'
-import { StyleSheet, View, Text } from "react-native"
+import { StyleSheet, View, Text, TouchableOpacity } from "react-native"
 
 export default function Settings({navigation}){
+
     return(
         <View style = {styles.container}>
             <Text
